Remove "--CREDITS--" from PHPT files
Most of those files have undergone massive changes; in some cases, the
"--CREDITS--" might not make sense.

This is aligned with the latest coding standard changes [1], since we no
longer allow the annotation "@author".

In any case, we can always use `git blame` or `git log` to see the
contributors.

[1]: 9a13c9fb03bebaaf6d79bf1b908f3c5c27918903

Signed-off-by: Henrique Moody <henriquemoody@gmail.com>