Run PHPStan in PHPT files
We were not running PHPStan on those files because of the strict types
declaration. However, I decided it would be better not to have strict
types on those files than not to run PHPStan issues.

Not surprisingly, there were several issues. Some issues were impacting
how user-friendly the library is.