Allow templates to be callables
There are a few use cases in which you would like to have a custom
exception but, at the same time, reuse the message or exception that
validation might give you.

This commit creates a new feature that allows users to define a callable
that will generate an exception when it fails.