Allow `Min` and `Max` to wrap chained rules
This follows the same pattern as the `Length` rule.

I've also removed the separate named and standard templates from these
rules. I didn't see a good way to implement these in the new pattern,
and also felt the language of the old standard template was somewhat
difficult to read anyway.