Always display the input
The "NotBlank", "NotEmpty", and "NotUndef" rules do not display the
input in all cases and instead displays the string "The value". The
problem with that is that one doesn't see which value was passed, which
is not so useful.

This commit will changes those rules to always display the input. If
someone doesn't want the input to show, they can always set a name for
the rule.