Do not use `array_unique()` in the `Unique` rule
The default value of the second argument of `array_unique()` is
`SORT_STRING`. This means that it converts all values to strings. That
was the reason we were using `SORT_REGULAR`, but when using `SORT_REGULAR`,
the strings `14.1` and `14.10` are considered the same, even though
they're different.

This commit will change the `Unique` rule to implement its own way of
detecting uniqueness. This was necessary because `array_unique() 's
limitation is handling different types in the same array.

Co-authored-by: Henrique Moody <henriquemoody@gmail.com>