Rename `StandardStringifier` to `ValidationStringifier`
The name `StandardStringifier` is ambiguous because it doesn't describe
how that stringifier differs from others. Ideally, we'd have a better
name for it, something that makes its behaviour explicit, but since this
is the stringifier from Validation, naming it as such will make it
easier for users to use it if they want.