Register `v` as a class alias to `Validator`
This is a pattern that we already use everywhere, so it makes sense to
simply turn `v` into a class alias to `Validator`, which will allow
people to use all the examples in the documentation without needing to
import the namespace.

We already do that when we run tests, so I just changed it to be a
global setting.