Create `All` rule and `all` prefix
This rule behaves similarly to ``Each, but the difference is that it
presents the messages in a simpler manner. This allows users to have a
simpler message when validating all items in an array or iterable
value.

I’m also introducing the `all` prefix, making it easier for users to
apply the `All` rule in combination with other rules.