Refactor the `Factory` class
The class was called `Factory` because we used to use it to create rules
and exceptions. Since we don’t have more exceptions per rule, the name
`RuleFactory` makes more sense, as it accurately describes its purpose.

I’ve turned that class into an interface, and copied the implementation
to a new class, `NamespacedRuleFactory`, which describes more of what
that specific implementation does.